remove hook namespace, move the hooks with the manager
